Code should be in a good/workable state.
